Rename impl module to internal and move entrypoint to cmd
