--- app/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/echo-webkom/cenv/cmd"
-
-func main() {
-	cmd.Run()
-}
